Card Class
Has all assets listed in the whiteboard drawings as well as getter and setter functions. Returns and gets vectors here and there so just be weary what you are passing it. For example, if you are replacing paths in a card, you have to make a vector with size 4 filled with the new values to pass it. blah blah blah you get the point...